requirements.txt
black==22.1.0
Bootstrap-Flask==2.0.1
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
install==1.3.5
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
python-dotenv==0.19.2
SQLAlchemy==1.4.31
tomli==2.0.1
typing-extensions==4.1.1
Werkzeug==2.0.2
WTForms==3.0.1
